planventure-api/requirements.txt
# Core dependencies
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0

# JWT and Authentication
flask-jwt-extended==4.4.4
PyJWT==2.8.0

# Data validation and serialization
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1

# Database and ORM
SQLAlchemy==2.0.20

# Development tools
python-dateutil==2.8.2